collector/pyproject.toml
[tool.poetry]
name = "collector"
version = "0.1.0"
description = ""
authors = ["Mikael Ivarsson <mikael.ivarsson4@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.10.7"
python-dotenv = "^1.0.0"
influxdb-client = "^1.36.1"
influxdb = "^5.3.1"
pymongo = {extras = ["srv"], version = "^4.3.3"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
